requirements.txt
asgiref==3.3.1
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.3.1
defusedxml==0.6.0
Django==3.1.4
django-admin-thumbnails==0.2.5
django-allauth==0.43.0
django-braces==1.14.0
django-ckeditor==6.0.0
django-contrib-comments==1.9.2
django-crispy-forms==1.10.0
django-debug-toolbar==3.2
django-etc==1.3.0
django-js-asset==1.2.2
django-markdown2==0.3.1
django-model-utils==4.0.0
django-mptt==0.11.0
django-polymorphic==3.0.0
django-polymorphic-tree==1.5.1
django-rest-auth==0.9.5
django-rest-swagger==2.2.0
django-star-ratings==0.9.1
django-tag-parser==3.2
django-taggit==1.2.0
djangorestframework==3.12.2
docutils==0.16
future==0.18.2
idna==2.10
install==1.3.4
itypes==1.2.0
Jinja2==2.11.2
Markdown==3.2.1
markdown2==2.3.10
MarkupSafe==1.1.1
oauthlib==3.1.0
openapi-codec==1.3.2
Pillow==8.0.1
psycopg2==2.8.6
pycparser==2.20
PyJWT==2.0.1
python3-openid==3.2.0
pytz==2020.5
PyYAML==5.3.1
requests==2.25.1
requests-oauthlib==1.3.0
simplejson==3.17.2
six==1.15.0
sqlparse==0.4.1
swapper==1.1.2.post1
uritemplate==3.0.1
urllib3==1.26.3
whitenoise==5.2.0
